adding EasyPackage 1.1.2 (Beta)
EasyPackage ist a tiny script which runs apt commands.
We use this because we wanted to make the usage of APT in our system a bit easyer. We know some people might dont like it but we do. If you are user of LGOS and you dont like EP, then you can easily remove it using "ep -remove-ep"